+++ stdlib/src/commonMain/kotlin/com/sunnychung/lib/multiplatform/kotlite/stdlib/AllStdLibModules.kt
@@ -1,5 +1,7 @@
 package com.sunnychung.lib.multiplatform.kotlite.stdlib
 
+import com.sunnychung.lib.multiplatform.kotlite.model.MultipleLibModules
+
 class AllStdLibModules(outputToConsoleFunction: (String) -> Unit = { print(it) }) : MultipleLibModules(
     name = "AllStdlib",
     modules = listOf(